@@ -1,27 +1,24 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 31
-    buildToolsVersion "31.0.0"
-
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
+    namespace "tun.proxy"
+    compileSdk 34
 
     defaultConfig {
         applicationId "tun.proxy"
         minSdkVersion 21
-        targetSdkVersion 31
-        versionCode 100270
-        versionName VERSION_NAME
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 34
+        versionCode 100209
+        versionName "1.2.9"
+
         externalNativeBuild {
             cmake {
                 cppFlags "-std=c++17 -fvisibility=hidden "
-                abiFilters 'armeabi-v7a','arm64-v8a', 'x86'
+                abiFilters 'armeabi-v7a','arm64-v8a', 'x86', 'x86_64'
                 arguments "-DCMAKE_VERBOSE_MAKEFILE=1 -DANDROID_FUNCTION_LEVEL_LINKING=ON"
             }
         }
+
         splits {
             abi {
                 enable true
