@@ -2,6 +2,7 @@ package tun.proxy;
 
 import android.app.Application;
 import android.content.SharedPreferences;
+
 import androidx.preference.PreferenceManager;
 
 import java.util.HashSet;
@@ -9,7 +10,7 @@ import java.util.Set;
 
 public class MyApplication extends Application {
     private final static String PREF_VPN_MODE = "pref_vpn_connection_mode";
-    private final static String PREF_APP_KEY[] = {"pref_vpn_disallowed_application", "pref_vpn_allowed_application"};
+    private final static String[] PREF_APP_KEY = {"pref_vpn_disallowed_application", "pref_vpn_allowed_application"};
 
     private static MyApplication instance;
 
@@ -23,11 +24,6 @@ public class MyApplication extends Application {
         instance = this;
     }
 
-    public enum VPNMode {DISALLOW, ALLOW};
-    public enum AppSortBy {APPNAME, PKGNAME};
-    public enum AppOrderBy {ASC, DESC};
-    public enum AppFiltertBy {APPNAME, PKGNAME};
-
     public VPNMode loadVPNMode() {
         final SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
         final String vpn_mode = sharedPreferences.getString(PREF_VPN_MODE, MyApplication.VPNMode.DISALLOW.name());
@@ -38,20 +34,23 @@ public class MyApplication extends Application {
         final SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
         final SharedPreferences.Editor editor = prefs.edit();
         editor.putString(PREF_VPN_MODE, mode.name()).apply();
-        return;
     }
 
     public Set<String> loadVPNApplication(VPNMode mode) {
         final SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
-        final Set<String> preference = prefs.getStringSet(PREF_APP_KEY[mode.ordinal()], new HashSet<String>());
-        return preference;
+        return prefs.getStringSet(PREF_APP_KEY[mode.ordinal()], new HashSet<>());
     }
 
     public void storeVPNApplication(VPNMode mode, final Set<String> set) {
         final SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
         final SharedPreferences.Editor editor = prefs.edit();
         editor.putStringSet(PREF_APP_KEY[mode.ordinal()], set).apply();
-        return;
     }
+
+    public enum VPNMode {DISALLOW, ALLOW}
+
+    public enum AppSortBy {APPNAME, PKGNAME}
+
+    public enum AppOrderBy {ASC, DESC}
 
 }
